requirements.txt
amqp==5.1.1
arabic-reshaper==3.0.0
asgiref==3.5.2
asn1crypto==1.5.1
async-timeout==4.0.2
autopep8==2.0.1
billiard==3.6.4.0
black==22.12.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
cssselect2==0.7.0
defusedxml==0.7.1
dj-database-url==1.2.0
Django==4.1.3
django-cors-headers==3.14.0
django-extensions==3.2.1
django-heroku==0.3.1
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
et-xmlfile==1.1.0
graphviz==0.20.1
gunicorn==20.1.0
html5lib==1.1
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
lxml==4.9.2
MarkupSafe==2.1.2
mypy-extensions==0.4.3
numpy==1.24.2
oauthlib==3.2.2
openpyxl==3.1.2
oscrypto==1.3.0
pandas==1.5.3
pathspec==0.10.3
Pillow==9.3.0
platformdirs==2.6.2
prompt-toolkit==3.0.38
psycopg2==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pydot==1.4.2
pyHanko==0.17.0
pyhanko-certvalidator==0.20.1
PyJWT==2.6.0
pyparsing==3.0.9
pypdf==3.4.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
redis==4.5.1
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.1
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
tomli==2.0.1
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
whitenoise==6.2.0
xhtml2pdf==0.2.9
